feat(ui-dioxus,gallery): showcase audit + Combobox/RadioGroup/DropdownMenu + Playwright wave
Spec: docs/superpowers/specs/2026-05-24-gallery-showcase-and-playwright-verification-design.md
Plan: docs/superpowers/plans/2026-05-24-gallery-showcase-and-playwright-verification.md

## New components (ui-dioxus)

- Combobox — typeahead-filtered single-select on Popover; pure
  `filter_options` helper, inline unit tests.
- RadioGroup — native `<input type=radio>` mutex with description copy,
  added to forms.rs.
- DropdownMenu — `role="menu"` overlay on Popover; first-class separator
  items.

Each (plus Select, DatePicker) exposes an additive, opt-in
`default_open: bool` prop so previews + SSR screenshots render the
overlay open without programmatic clicks.

## Gallery showcase remediation

Previews now visibly demonstrate documented capabilities, not just
instantiate the component:

- inputs.rs: TextField 3-tile grid (default + invalid + disabled);
  Select / DatePicker / Combobox open by default; RadioGroup added.
- actions.rs: DropdownMenu open by default.
- feedback.rs: Dialog open by default; Popover open; Tooltip 2-tile
  grid (always-visible showcase + hover/focus); Toast pre-seeded with
  one of each tone so the static surface always renders.
- surfaces.rs: MetricCard 4-tone grid. GlassSurface left single-tile
  on purpose — multiple instances trigger webkit's WebGL context cap.

## Playwright e2e

- New behavioral spec files for the three new components.
- Refreshed dialog/toast/tooltip/glass-surface specs for the new
  preview shapes.
- Universal canvas mask in visual.spec.ts (LiquidSurface AND
  GlassSurface render canvases on WgpuWebGl2 tier).
- 15s screenshot timeout so unoptimized wasm (wasm-opt fails on this
  Windows machine with 0xc0000409) still completes first-paint in
  time.
- component-manifest.ts: registered Combobox, RadioGroup, DropdownMenu.
- Fresh visual baselines for the 12 components whose previews changed.

## playwright.config.ts bug fix

`DIST_DIR` was unquoted in the http-server shell command — Windows
paths containing spaces broke the lookup and served 404s. Quoting
restores local Windows e2e.

## Workspace lint cleanup

Cleared 4 originally-flagged ui-glass-engine clippy errors + cascade
(manual_div_ceil ×6, is_multiple_of, cloned_ref_to_slice_refs,
too_many_arguments, unnecessary_cast). Gated
ui-glass-engine integration tests behind `required-features =
["headless"]` so workspace clippy skips them when the feature is off
(matches the existing `#[cfg(feature = "headless")]` gate in
`src/lib.rs`). Wasm32 clippy clean (thread_local const init, RAF
closure type alias, file-scoped allows for non-Send/Sync wgpu Arc).

## Tests

- Cargo workspace: 289 tests pass (excluding ui-glass-engine GPU
  integration tests, which require the headless feature).
- Cargo fmt + clippy (native + wasm32) clean.
- Playwright (filtered to the 12 affected components): 241/242 passing
  on first run; GlassSurface visual webkit verified separately (8/8).
- Pre-existing visual-baseline drift remains for 33 unchanged
  components against this Windows machine. The CI workflow runs on
  Linux and generates its own `-linux.png` baselines as artifacts, so
  this drift does not affect the CI green path.

+++ crates/ui-glass-engine/src/pipeline.rs
@@ -65,8 +65,8 @@ pub fn build_blur_pipeline(
     });
 
     let constants: &[(&str, f64)] = &[
-        ("BLUR_DIRECTION_X", dx as f64),
-        ("BLUR_DIRECTION_Y", dy as f64),
+        ("BLUR_DIRECTION_X", dx),
+        ("BLUR_DIRECTION_Y", dy),
         ("BLUR_TAPS", taps as f64),
     ];
